Restrict the surface filtering to only footpaths
This commit makes the surfaces value only be stored for ways which are
footpaths. Before this change, 3 058 238 surface tags were processed.
After it, only 602700 surface tags were saved. I think it would be a
good argument for correctness, but after discussion with Christian, we
decided that it is too much of a niche case and best to leave it simple
for now. I leave this on a branch in case it is desired in the future.